♻️ Refacotr: util 함수 분리
refacotr: util 함수 분리
